+++ posthog-server/CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - feat: `timestamp` can now be overridden when capturing an event ([#297](https://github.com/PostHog/posthog-android/issues/297))
 - feat: Add `groups`, `groupProperties`, `personProperties` overrides to feature flag methods ([#298](https://github.com/PostHog/posthog-android/issues/298))
+- feat: Add local evaluation for feature flags
 
 ## 1.0.3 - 2025-10-01